requirements.txt
fastapi
uvicorn
google-cloud-dialogflow